Move assert_match to the tests as is only used there
